backend/requirements.txt
python-dotenv
newsapi-python
praw
pymongo
pydantic